Delete and Update now working
Bank Model:
Added setValueAt() which takes an input and sets it to the
corresponding category in the selected account. Changed update() and
delete() to accept Integers instead of account types. Also they now
will only delete the row selected instead of many rows.

BankGUI:
Added helper methods for setting the GregorianCalendar (thinking we can
tweak for other calendar and it will stay helpful) and for finding the
index of the row selected. Added functionality to jbtUpdate and
jbtDelete in ButtonListener class.